+++ ssc/interpolation/PiecewiseConstant.cpp
@@ -27,7 +27,7 @@ ConstantStepInterpolator(xmin_,xmax_,y_)
 
 void PiecewiseConstant::update_coefficients_if_necessary(const double val)
 {
-    idx = std::max(0,(int)std::min(floor((val-xmin)/(xmax-xmin)*double(n)),(double)n-1));
+    idx = (size_t)std::max(0,(int)std::min(floor((val-xmin)/(xmax-xmin)*double(n)),(double)n-1));
 }
 
 void PiecewiseConstant::compute_coefficients_for_ith_interval(const double x0, const size_t i)
