+++ ssc/numeric/kahan_sum.hpp
@@ -51,9 +51,9 @@ namespace ssc
 
         template <typename Iterator, typename T>  T kahan(const Iterator& begin, const Iterator& end)
         {
-            return kahan<typename boost::counting_iterator<size_t>,
+            return kahan<typename boost::counting_iterator<long>,
                              Iterator,
-                             T>(boost::counting_iterator<size_t>(0),boost::counting_iterator<size_t>(std::distance(begin,end)), begin);
+                             T>(boost::counting_iterator<long>(0),boost::counting_iterator<long>(std::distance(begin,end)), begin);
         }
 
         template <typename T>  T kahan(const std::vector<T>& points)
